requirements.txt
# Core Framework
# Keep crewai at version 0.130.0
# Only change crewai version if there are critical dependency conflicts
crewai==0.130.0
crewai-tools==0.47.1

# Web Framework
fastapi==0.110.3
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# LLM Provider
litellm>=1.40.0

# PDF Processing
langchain-community>=0.0.38
pypdf>=4.0.0

# Environment
python-dotenv>=1.0.0

# Google AI (optional - for Google Gemini)
google-generativeai>=0.5.4

# Data Processing
pandas>=2.2.2
numpy>=1.26.4
pydantic>=2.0.0

# Search Tool
google-search-results>=2.4.2

# Database (Bonus)
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Queue Worker (Bonus)
celery[redis]>=5.3.0
redis>=5.0.0